refactor(test): move mock factory to mocks/ directory per coding standards
Move createMockRewriteRuleRepository.ts to mocks/ subdirectory following
the mock placement rules in basic-rule.md. Update all import paths in
test files and documentation.

Also add a list of non-compliant mock factories to basic-rule.md for
future refactoring reference.
