+++ host-frontend-root/frontend-src-root/tests/unit/application-business-rules/interactors/ToggleRuleActiveInteractor/execute/error-cases.test.ts
@@ -3,7 +3,7 @@
  * 1. repository.getByIdでエラーが発生した場合、presentErrorが呼び出される
  * 2. repository.updateでエラーが発生した場合、presentErrorが呼び出される
  */
-import { createMockRewriteRuleRepository as createMockRepository } from 'tests/unit/application/ports/IRewriteRuleRepository/createMockRewriteRuleRepository';
+import { createMockRewriteRuleRepository as createMockRepository } from 'tests/unit/application/ports/IRewriteRuleRepository/mocks/createMockRewriteRuleRepository';
 import { createMockPresenter } from 'tests/unit/application-business-rules/interactors/ToggleRuleActiveInteractor/mocks/createMockPresenter';
 import { createMockTabsGateway } from 'tests/unit/application-business-rules/interactors/ToggleRuleActiveInteractor/mocks/createMockTabsGateway';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
