+++ host-frontend-root/frontend-src-root/tests/unit/application/ports/IRewriteRuleRepository/mocks/createMockRewriteRuleRepository.ts
